Fixed behavior of demo; however the book seems to rely on some non-obvious behavior. Notes follow:
I attempted to use a vertex buffer in place of the book's call to
glVertexAttribI1i. This fits the framework I'm starting to build a bit
better. However, I think this couldn't work (after a lot of frustration)
because the book generates the vertices within the vertex shader. So, at
the call to glDrawArrays there's never any call to index into the buffer
(because there are no vertices to index over). While I had the buffer
bound, my framework causes the shader attribute it's linked to (at
layout(location = 0) to be enabled. This seems to "block" the call to
force data through via glVertexAttribI1i. After removing my vertex
buffer completely and disabling the droplet_index variable (as inspected
via renderdoc) then calls to glVertexAttribI1i actually pass data and
the sample works.